fix(lower): signed arithmetic wraps, as Go says it does
Go defines signed integer overflow as wrapping. C leaves it undefined, which is
not a wrong number but a licence for the compiler to assume it cannot happen --
so the generated kernel had no defined meaning on exactly the values the Go
source did define, it compiled without a word, and what it did depended on the
optimiser. The differential fuzzer found it on `11 - (x << 63)`, where the shift
is MinInt64 and the subtraction overflows; the host run and the emulator then
disagreed by whole powers of two on every element of the buffer.

`+`, `-`, `*`, unary `-` and `<<` on a signed integer are now emitted through
the unsigned type of the same width, which is the only modular arithmetic C
defines. The shift already was; this makes it one rule rather than a special
case, and signedShiftLeft is gone.

The whole difficulty was doing it without burying the source. Converting at
every operator would turn `a + b*c` into a cast four deep. So a *region* of
wrapping arithmetic is converted once at its boundary instead:

	(int)((unsigned int)(a) + (unsigned int)(b) * (unsigned int)(c))

`/`, `%` and `>>` end a region rather than joining it, because they mean
something different on unsigned operands. `&`, `|` and `^` are bit-identical at
the same width and would be safe either way; they are left outside too, so the
rule has no exception to remember. A shift count is not a term of the
arithmetic, so it stays outside as well.

The readability cost is real and was the accepted trade: `h[i+1]` becomes
`h[(int)((unsigned int)(i) + 1u)]`. The cost in anything else is close to
nothing. Six of the twelve kernels changed, and of their PTX **three are
byte-identical** -- FIR, Quantize and Transpose -- Gray is the same size,
BandGain grows 28 bytes, and Classify shrinks by 483. Taking away an assumption
the optimiser was entitled to make did not cost instructions; in one case it
saved them.

Two limits stay, and SPEC.md states both rather than leaving them to be found.
This makes the C defined, not equal to Go, for Go's int: it wraps at 64 bits
where the device wraps at 32, which is the narrowing SPEC.md already calls the
one deliberate infidelity. And MinInt / -1 is still undefined on the device,
which routing through unsigned cannot fix, unsigned division being a different
operation.

Every golden and every prebuilt artifact was regenerated by its own tooling and
the diffs read, not blessed: the load-bearing one is Gray's
`77*r + ... + 128 >> 8`, which becomes `(int)(77u*... + 128u) >> 8` -- the same
meaning, now explicit, with the shift correctly outside the region.

The failing input is committed as a corpus entry, this time only after it
passed. It fails against the emitter without this change.

+++ kernels/prebuilt/Classify.cu
@@ -10,7 +10,7 @@ extern "C" __global__ void Classify(float* __restrict__ out, int out_len, const 
 	int band = edges_len;
 	for (int base = 0; base < edges_len; base += 4)
 	{
-		for (int j = base; j < base + 4; j++)
+		for (int j = base; j < (int)((unsigned int)(base) + 4u); j++)
 		{
 			if (j >= edges_len)
 			{
